Move api calls to api/
