+++ src/api/fetchFiveRandomAlbums.js
@@ -1,4 +1,4 @@
-import { wordBank } from './words';
+import { wordBank } from '../components/words';
 
 const baseURL = process.env.REACT_APP_SPOTIFY_BASE_URL;
