+++ src/components/prepareFiveRoundsData.js
@@ -1,6 +1,6 @@
-import { fetchFiveRandomAlbums } from './fetchFiveRandomAlbums';
-import { fetchArtistData } from './fetchArtistData';
-import { fetchRelatedArtists } from './fetchRelatedArtists';
+import { fetchFiveRandomAlbums } from '../api/fetchFiveRandomAlbums';
+import { fetchArtistData } from '../api/fetchArtistData';
+import { fetchRelatedArtists } from '../api/fetchRelatedArtists';
 import { getRoundArtistId } from './getRoundArtistId';
 import { extractAlbumData } from './extractAlbumData';
 import { extractArtistData } from './extractArtistData';
